collect into refactoring for using
