+++ src/collect_into/mod.rs
@@ -4,6 +4,4 @@ mod split_vec;
 mod vec;
 
 pub use par_collect_into::ParCollectInto;
-
-#[cfg(test)]
-pub use par_collect_into::ParCollectIntoCore;
+pub(crate) use par_collect_into::ParCollectIntoCore;
