+++ src/collect_into/par_collect_into.rs
@@ -1,7 +1,6 @@
 use crate::computations::{M, Values, X, Xfx};
 use crate::runner::ParallelRunner;
-use crate::using::Using;
-use crate::using::computations::{UM, UX, UXfx};
+use crate::using::UParCollectIntoCore;
 use orx_concurrent_iter::ConcurrentIter;
 use orx_iterable::Collection;
 use orx_pinned_vec::IntoConcurrentPinnedVec;
@@ -35,37 +34,6 @@ pub trait ParCollectIntoCore<O: Send + Sync>: Collection<Item = O> {
         M1: Fn(I::Item) -> Vt + Send + Sync,
         F: Fn(&Vt::Item) -> bool + Send + Sync,
         M2: Fn(Vt::Item) -> Vo + Send + Sync;
-
-    fn u_m_collect_into<R, U, I, M1>(self, m: UM<U, I, O, M1>) -> Self
-    where
-        R: ParallelRunner,
-        U: Using,
-        I: ConcurrentIter,
-        M1: Fn(&mut U::Item, I::Item) -> O + Send + Sync;
-
-    fn u_x_collect_into<R, U, I, Vo, M1>(self, x: UX<U, I, Vo, M1>) -> Self
-    where
-        R: ParallelRunner,
-        U: Using,
-        I: ConcurrentIter,
-        Vo: Values<Item = O> + Send + Sync,
-        Vo::Item: Send + Sync,
-        M1: Fn(&mut U::Item, I::Item) -> Vo + Send + Sync;
-
-    fn u_xfx_collect_into<R, U, I, Vt, Vo, M1, F, M2>(
-        self,
-        xfx: UXfx<U, I, Vt, Vo, M1, F, M2>,
-    ) -> Self
-    where
-        R: ParallelRunner,
-        U: Using,
-        I: ConcurrentIter,
-        Vt: Values + Send + Sync,
-        Vt::Item: Send + Sync,
-        Vo: Values<Item = O> + Send + Sync,
-        M1: Fn(&mut U::Item, I::Item) -> Vt + Send + Sync,
-        F: Fn(&mut U::Item, &Vt::Item) -> bool + Send + Sync,
-        M2: Fn(&mut U::Item, Vt::Item) -> Vo + Send + Sync;
 
     // test
 
@@ -113,6 +81,9 @@ pub trait ParCollectIntoCore<O: Send + Sync>: Collection<Item = O> {
 }
 
 /// Collection types into which outputs of a parallel computations can be collected into.
-pub trait ParCollectInto<O: Send + Sync>: ParCollectIntoCore<O> {}
+pub trait ParCollectInto<O: Send + Sync>: ParCollectIntoCore<O> + UParCollectIntoCore<O> {}
 
-impl<O: Send + Sync, C: ParCollectIntoCore<O>> ParCollectInto<O> for C {}
+impl<O: Send + Sync, C> ParCollectInto<O> for C where
+    C: ParCollectIntoCore<O> + UParCollectIntoCore<O>
+{
+}
