+++ src/using/collect_into/mod.rs
@@ -1,0 +1,6 @@
+mod fixed_vec;
+mod split_vec;
+mod u_par_collect_into;
+mod vec;
+
+pub use u_par_collect_into::UParCollectIntoCore;
