+++ src/using/collect_into/vec.rs
@@ -1,34 +1,28 @@
-use super::par_collect_into::ParCollectIntoCore;
-use crate::computations::{M, Values, X, Xfx};
+use crate::computations::Values;
 use crate::runner::ParallelRunner;
+use crate::using::Using;
+use crate::using::collect_into::u_par_collect_into::UParCollectIntoCore;
+use crate::using::computations::{UM, UX, UXfx};
 use orx_concurrent_iter::ConcurrentIter;
 use orx_fixed_vec::FixedVec;
 use orx_pinned_vec::PinnedVec;
 use orx_split_vec::{GrowthWithConstantTimeAccess, SplitVec};
 
-impl<O> ParCollectIntoCore<O> for Vec<O>
+impl<O> UParCollectIntoCore<O> for Vec<O>
 where
     O: Send + Sync,
 {
-    type BridgePinnedVec = FixedVec<O>;
-
-    fn empty(iter_len: Option<usize>) -> Self {
-        match iter_len {
-            Some(len) => Vec::with_capacity(len),
-            None => Vec::new(),
-        }
-    }
-
-    fn m_collect_into<R, I, M1>(mut self, m: M<I, O, M1>) -> Self
+    fn u_m_collect_into<R, U, I, M1>(mut self, m: UM<U, I, O, M1>) -> Self
     where
         R: ParallelRunner,
+        U: Using,
         I: ConcurrentIter,
-        M1: Fn(I::Item) -> O + Send + Sync,
+        M1: Fn(&mut U::Item, I::Item) -> O + Send + Sync,
     {
         match m.par_len() {
             None => {
                 let split_vec = SplitVec::with_doubling_growth_and_max_concurrent_capacity();
-                let split_vec = split_vec.m_collect_into::<R, _, _>(m);
+                let split_vec = split_vec.u_m_collect_into::<R, _, _, _>(m);
                 extend_from_split(self, split_vec)
             }
             Some(len) => {
@@ -40,40 +34,38 @@ where
         }
     }
 
-    fn x_collect_into<R, I, Vo, M1>(self, x: X<I, Vo, M1>) -> Self
+    fn u_x_collect_into<R, U, I, Vo, M1>(self, x: UX<U, I, Vo, M1>) -> Self
     where
         R: ParallelRunner,
+        U: Using,
         I: ConcurrentIter,
         Vo: Values<Item = O> + Send + Sync,
         Vo::Item: Send + Sync,
-        M1: Fn(I::Item) -> Vo + Send + Sync,
+        M1: Fn(&mut U::Item, I::Item) -> Vo + Send + Sync,
     {
         let split_vec = SplitVec::with_doubling_growth_and_max_concurrent_capacity();
-        let split_vec = split_vec.x_collect_into::<R, _, _, _>(x);
+        let split_vec = split_vec.u_x_collect_into::<R, _, _, _, _>(x);
         extend_from_split(self, split_vec)
     }
 
-    fn xfx_collect_into<R, I, Vt, Vo, M1, F, M2>(self, xfx: Xfx<I, Vt, Vo, M1, F, M2>) -> Self
+    fn u_xfx_collect_into<R, U, I, Vt, Vo, M1, F, M2>(
+        self,
+        xfx: UXfx<U, I, Vt, Vo, M1, F, M2>,
+    ) -> Self
     where
         R: ParallelRunner,
+        U: Using,
         I: ConcurrentIter,
         Vt: Values + Send + Sync,
         Vt::Item: Send + Sync,
         Vo: Values<Item = O> + Send + Sync,
-        M1: Fn(I::Item) -> Vt + Send + Sync,
-        F: Fn(&Vt::Item) -> bool + Send + Sync,
-        M2: Fn(Vt::Item) -> Vo + Send + Sync,
+        M1: Fn(&mut U::Item, I::Item) -> Vt + Send + Sync,
+        F: Fn(&mut U::Item, &Vt::Item) -> bool + Send + Sync,
+        M2: Fn(&mut U::Item, Vt::Item) -> Vo + Send + Sync,
     {
         let split_vec = SplitVec::with_doubling_growth_and_max_concurrent_capacity();
-        let split_vec = split_vec.xfx_collect_into::<R, _, _, _, _, _, _>(xfx);
+        let split_vec = split_vec.u_xfx_collect_into::<R, _, _, _, _, _, _, _>(xfx);
         extend_from_split(self, split_vec)
-    }
-
-    // test
-
-    #[cfg(test)]
-    fn length(&self) -> usize {
-        self.len()
     }
 }
