+++ src/using/mod.rs
@@ -1,8 +1,10 @@
+mod collect_into;
 pub mod computational_variants;
-pub(crate) mod computations;
+mod computations;
 mod runner;
 mod u_par_iter;
 mod using_variants;
 
+pub(crate) use collect_into::UParCollectIntoCore;
 pub use u_par_iter::ParIterUsing;
 pub use using_variants::{Using, UsingClone, UsingFun};
